Enable renovate-test and renovate_test_2
renovate-test and renovate_test_2 were moved to a new namespace.
They used to be automatically selected by renovate-runner in
autodiscover mode since they were part of the
platform-engineering-org namespace. Since we moved them to a new
namespace we now need to explicitly define them in the configuration
for renovate-runner to be able to detect them.

Signed-off-by: Giulia Naponiello <gnaponie@redhat.com>